+++ apps/core/services/entities/__init__.py
@@ -1,0 +1,13 @@
+from .artwork_basic_info import ArtworkBasicInfo
+from .extraction_metadata import ExtractionMetadata
+from .web_search_info import WebSearchInfo
+from .content_fetch_info import ContentFetchInfo
+from .artwork_extracted_info import ArtworkExtractedInfo
+
+__all__ = [
+    'ArtworkBasicInfo',
+    'ExtractionMetadata',
+    'WebSearchInfo',
+    'ContentFetchInfo',
+    'ArtworkExtractedInfo',
+]
