+++ apps/core/services/externals/__init__.py
@@ -1,0 +1,9 @@
+from .gemini_service import GeminiService
+from .brave_service import BraveSearchService
+from .fetch_service import FetchService
+
+__all__ = [
+    'GeminiService',
+    'BraveSearchService',
+    'FetchService',
+]
