+++ apps/core/services/externals/fetch_service.py
