+++ apps/core/services/externals/gemini_service.py
