+++ pyproject.toml
@@ -19,6 +19,7 @@ dependencies = [
     "google-auth>=2.40.3",
     "google-cloud-aiplatform>=1.104.0",
     "vertexai>=1.43.0",
+    "python-decouple>=3.8",
 ]
 
 [project.optional-dependencies]
